* move some extension to file
