+++ DarkEggKit/Classes/AniImageView/AImageView+Animator.swift
@@ -255,9 +255,3 @@ extension AImageView {
         }
     }
 }
-
-extension Array {
-    subscript(safe index: Int) -> Element? {
-        return indices ~= index ? self[index] : nil
-    }
-}
